requirements.txt
fastapi
uvicorn
email-validator
disposable-email-domains
requests
dnspython

